Create ToDoList#delete endpoint
