+++ app/validators/common/schemas/to_do_lists/destroy_to_do_list_schema.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Common
+  module Schemas
+    module ToDoLists
+      DestroyToDoListSchema = Dry::Schema.Params do
+        required(:id).filled(:string)
+      end
+    end
+  end
+end
